lmb-922: fractie selector html component now also checks of it needs to be required for the bestuursorgaan
