+++ app/components/rdf-input-fields/mandataris-fractie-selector.js
@@ -11,6 +11,7 @@ import { loadBestuursorgaanUrisFromContext } from 'frontend-lmb/utils/form-conte
 import { MANDAAT } from 'frontend-lmb/rdf/namespaces';
 import { isPredicateInObserverChange } from 'frontend-lmb/utils/is-predicate-in-observer-change';
 import { MANDATARIS_PREDICATE } from 'frontend-lmb/utils/constants';
+import { isRequiredForBestuursorgaan } from 'frontend-lmb/utils/is-fractie-selector-required';
 
 import { NamedNode } from 'rdflib';
 import { restartableTask, timeout } from 'ember-concurrency';
@@ -86,22 +87,8 @@ export default class MandatarisFractieSelector extends InputFieldComponent {
     )[0];
 
     this.bestuursperiode = await bestuursorgaan.heeftBestuursperiode;
-    await this.setIsRequiredForBestuursorgaan(bestuursorgaan);
-  }
-
-  async setIsRequiredForBestuursorgaan(bestuursorgaan) {
-    if (!bestuursorgaan) {
-      this.isRequiredForBestuursorgaan = this.isRequired;
-      return;
-    }
-
-    const requiredWhen = [
-      await bestuursorgaan.isGR,
-      await bestuursorgaan.isCBS,
-      await bestuursorgaan.isBurgemeester,
-    ];
-
-    this.isRequiredForBestuursorgaan = requiredWhen.some((is) => is === true);
+    this.isRequiredForBestuursorgaan =
+      await isRequiredForBestuursorgaan(bestuursorgaan);
   }
 
   async loadProvidedValue() {
